% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestep.R
\name{A.fun}
\alias{A.fun}
\title{Calculate negative information (A) for logistic regression.}
\usage{
A.fun(bet, dat)
}
\arguments{
\item{bet}{is the beta for evaluation}

\item{dat}{is the dataset}
}
\description{
\code{A.fun} evaluates negative information (A) for logistic regression.
}
