% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_wrappers.R
\name{Est.ALASSO.Approx.GLMNET}
\alias{Est.ALASSO.Approx.GLMNET}
\title{Least square approximated adaptive lasso}
\usage{
Est.ALASSO.Approx.GLMNET(ynew, xnew, bini, N.adj, lambda.grid, modBIC = T,
  N.inflate = N.adj)
}
\arguments{
\item{ynew}{a p by 1 vector of LSA-based outcome}

\item{xnew}{a p by p matrix of LSA-based covariates}

\item{bini}{1/bini is used as the penalty for adaptive lasso}

\item{N.adj}{log(N.adj) is used to penalize BIC, where BIC = -2loglik + df*log(N.adj), and modified BIC, where modBIC = -2loglik + df*N.adj^0.1}

\item{lambda.grid}{the grid of lambda to put into glmnet}

\item{modBIC}{if the program also finds the lamda that minimizes the modified BIC. The default is TRUE.}

\item{N.inflate}{the factor to inflate sum residual squared}
}
\value{
bhat.BIC adaptive lasso estimator using BIC and bhat.modBIC adaptive lasso estimator using modified BIC. lambda.BIC and lambda.modBIC return the optimal lambda chosen by BIC and modified BIC.
}
\description{
\code{Est.ALASSO.Approx.GLMNET} fits adaptive lasso based on least square approximation (LSA)
}
\examples{
Est.ALASSO.Approx.GLMNET(ynew,xnew,bini,N.adj)
}
\author{
Yan Wang, Tianxi Cai
}
