% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_wrappers.R
\name{Est.ALASSO.GLMNET.CV}
\alias{Est.ALASSO.GLMNET.CV}
\title{A modified wrapper for adaptive lasso with the optimal penalty chosen by cross-validation}
\usage{
Est.ALASSO.GLMNET.CV(data, fam0 = "binomial", w.b = NULL, lambda.grid,
  chunksize = 50)
}
\arguments{
\item{data}{matrix or data frame. For non-survival outcome, first column = response, rest = design matrix.
For continuous-time survival outcome, first column = time, second column = delta (0/1), rest = design matrix.}

\item{fam0}{family of the response, taking "binomial", "Poisson", "Cox"}

\item{w.b}{w.b used to penalize adaptive lasso. If null, a glm/Cox model will be fitted and 1/abs(coefficients) will be used as w.b}

\item{lambda.grid}{the grid of lambda to put into glmnet}

\item{chunksize}{The prediction step in cv.glmnet could take a large amount of memory. chunksize specifies how many chunks you would like to split the prediction step. The predition step will be run in a loop if chunksize>1.}
}
\value{
a list containing two arguments: bhat.cv adaptive lasso estimator using 10-fold cross-validation; lambda.cv is the optimal lambda chosen by cross-validation.
}
\description{
\code{Est.ALASSO.GLMNET.CV} fits adaptive lasso based on cv.glmnet. The best lambda (penalizing factor) is chosen by 10-fold cross-validation.
}
\examples{
Est.ALASSO.GLMNET.CV(data, fam0="binomial", w.b = NULL, lambda.grid, chunksize = 50)
}
\author{
Yan Wang, Tianxi Cai
}
