% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_wrappers.R
\name{Est.ALASSO.GLMNET}
\alias{Est.ALASSO.GLMNET}
\title{A wrapper for adaptive lasso}
\usage{
Est.ALASSO.GLMNET(data, BIC.factor = 0.1, fam0 = "binomial", w.b = NULL,
  lambda.grid, modBIC = T)
}
\arguments{
\item{data}{matrix or data frame. For non-survival outcome, first column = response, rest = design matrix.
For continuous-time survival outcome, first column = time, second column = delta (0/1), rest = design matrix.}

\item{BIC.factor}{factor in modified BIC, BIC = -2 loglikelihood + df * N^BIC.factor}

\item{fam0}{family of the response, taking "binomial", "Poisson", "Cox"}

\item{w.b}{w.b used to penalize adaptive lasso. If null, a glm/Cox model will be fitted and 1/abs(coefficients) will be used as w.b}

\item{lambda.grid}{the grid of lambda to put into glmnet}

\item{modBIC}{if the program also finds the lamda that minimizes the modified BIC. The default is TRUE.}
}
\value{
bhat.BIC adaptive lasso estimator using BIC and bhat.modBIC adaptive lasso estimator using modified BIC. lambda.BIC and lambda.modBIC return the optimal lambda chosen by BIC and modified BIC.
}
\description{
\code{Est.ALASSO.GLMNET} fits adaptive lasso based on glmnet. The best lambda (penalizing factor) is chosen by BIC or modified BIC.
}
\examples{
Est.ALASSO.Approx.GLMNET(ynew,xnew,bini,N.adj)
}
\author{
Yan Wang, Tianxi Cai
}
