% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_wrappers.R
\name{Est.ALASSO.GLMNET.TANGXIE}
\alias{Est.ALASSO.GLMNET.TANGXIE}
\title{A wrapper for divide-and-conquer adaptive lasso proposed by Chen and Xie (2014) and Tang et al. (2016).}
\usage{
Est.ALASSO.GLMNET.TANGXIE(dat.list, K, BIC.factor = 0.1, fam0 = "binomial",
  lambda.grid, mvpct = 0.5, modBIC = T, adaptive = T, onestep = F)
}
\arguments{
\item{dat.list}{a list of matrices. Each element of the list is a sub-dataset. In each sub-dataset, for non-survival outcome, first column = response, rest = design matrix.
In each sub-dataset, for continuous-time survival outcome, first column = time, second column = delta (0/1), rest = design matrix.}

\item{K}{Number of sub-datasets in dat.list}

\item{BIC.factor}{factor in modified BIC, BIC = -2 loglikelihood + df * N^BIC.factor}

\item{fam0}{family of the response, taking "binomial", "Poisson", "Cox"}

\item{lambda.grid}{the grid of lambda to put into glmnet}

\item{mvpct}{majority voting percentage used for Chen and Xie (2014)}

\item{modBIC}{if the program also finds the lamda that minimizes the modified BIC. The default is TRUE.}

\item{adaptive}{if adaptive lasso is used. The default is TRUE}

\item{onestep}{if one-step estimator should be used as the initial estimator for Cox fit}

\item{w.b}{w.b used to penalize adaptive lasso. If null, a glm/Cox model will be fitted and 1/abs(coefficients) will be used as w.b}
}
\value{
a list containing two arguments: bhat.cv adaptive lasso estimator using 10-fold cross-validation; lambda.cv is the optimal lambda chosen by cross-validation.
}
\description{
\code{Est.ALASSO.GLMNET.TANGXIE} fits adaptive lasso based on cv.glmnet. The best lambda (penalizing factor) is chosen by 10-fold cross-validation.
}
\examples{
Est.ALASSO.GLMNET.TANGXIE(dat.list,K,BIC.factor=0.1,fam0="binomial",lambda.grid,mvpct = 0.5)
}
\references{
Chen, Xueying, and Min-ge Xie. "A split-and-conquer approach for analysis of extraordinarily large data." Statistica Sinica (2014): 1655-1684.

Tang, Lu, Ling Zhou, and Peter X-K. Song. "Method of Divide-and-Combine in Regularised Generalised Linear Models for Big Data." arXiv preprint arXiv:1611.06208 (2016).
}
\author{
Yan Wang, Tianxi Cai
}
