% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestep.R
\name{iteration.fun.cox2}
\alias{iteration.fun.cox2}
\title{One-step iteration function for divide-and-conquer Cox model}
\usage{
iteration.fun.cox2(dat.list, bini, kk.list)
}
\arguments{
\item{dat.list}{list of subsets (after dividing). In each subset, first column = U, second column = delta (0/1), rest = design matrix}

\item{bini}{initial estimator as starting point}

\item{kk.list}{which subsets of dat.list get one-step update}
}
\value{
a list with b.k a matrix of one-step estimators and with Ahat the negative information matrix
}
\description{
\code{iteration.fun.cox2} estimates a one-step for a Cox regression for each subset. The core used a C function.
}
\examples{
iteration.fun.cox(dat.list=dat.list,bini=bini,kk.list=2:K,rtn='Score+A+Approx')
}
\author{
Yan Wang, Tianxi Cai
}
