% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestep.R
\name{logitlik.fun}
\alias{logitlik.fun}
\title{Calculate log-likelihood for logistic regression.}
\usage{
logitlik.fun(bet.mat, dat)
}
\arguments{
\item{bet.mat}{is the beta matrix for evaluation}

\item{dat}{is the dataset}
}
\description{
\code{logitlik.fun} evaluates log-likelihood for logistic regression.
}
