% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_wrappers.R
\name{MCP.Xie}
\alias{MCP.Xie}
\title{A wrapper for divide-and-conquer MCP Cox model proposed by Chen and Xie (2014) .}
\usage{
MCP.Xie(dat.list, K, lambda.grid, mvpct = 0.5)
}
\arguments{
\item{dat.list}{a list of matrices. Each element of the list is a sub-dataset. In each sub-dataset, for non-survival outcome, first column = response, rest = design matrix.
In each sub-dataset, for continuous-time survival outcome, first column = time, second column = delta (0/1), rest = design matrix.}

\item{K}{Number of sub-datasets in dat.list}

\item{lambda.grid}{the grid of lambda to put into glmnet}

\item{mvpct}{majority voting percentage used for Chen and Xie (2014)}

\item{BIC.factor}{factor in modified BIC, BIC = -2 loglikelihood + df * N^BIC.factor}
}
\value{
a list containing the vector of estimates.
}
\description{
\code{MCP.Xie} fits Cox model with an MCP penalty. The best lambda (penalizing factor) is chosen by BIC.
}
\examples{
MCP.Xie(dat.list,K,BIC.factor=0.1,lambda.grid,mvpct = 0.5)
}
\references{
Chen, Xueying, and Min-ge Xie. "A split-and-conquer approach for analysis of extraordinarily large data." Statistica Sinica (2014): 1655-1684.
}
\author{
Yan Wang, Tianxi Cai
}
