% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvglmnet_util.R
\name{mycv.coxnet}
\alias{mycv.coxnet}
\title{Inner function for Est.ALASSO.GLMNET.CV}
\usage{
mycv.coxnet(outlist, lambda, x, y, weights, offset, foldid, type.measure,
  grouped, keep = FALSE)
}
\description{
\code{mycv.coxnet} is a modified function of cv.coxnet, such that the prediction step can be split and run.
}
