% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvglmnet_util.R
\name{mycv.glmnet}
\alias{mycv.glmnet}
\title{Inner function for Est.ALASSO.GLMNET.CV}
\usage{
mycv.glmnet(x, y, weights, offset = NULL, lambda = NULL,
  type.measure = c("mse", "deviance", "class", "auc", "mae"), nfolds = 10,
  foldid, grouped = TRUE, keep = FALSE, parallel = FALSE, ...)
}
\description{
\code{mycv.glmnet} is a modified function of cv.glmnet, such that the prediction step can be split and run.
}
