% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvglmnet_util.R
\name{mycv.lognet}
\alias{mycv.lognet}
\title{Inner function for Est.ALASSO.GLMNET.CV}
\usage{
mycv.lognet(outlist, lambda, x, y, weights, offset, foldid, type.measure,
  grouped, keep = FALSE)
}
\description{
\code{mycv.lognet} is a modified function of cv.lognet, such that the prediction step can be split and run.
}
