% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpxie_util.R
\name{ncvsurv2}
\alias{ncvsurv2}
\title{Inner function for MCP.Xie}
\usage{
ncvsurv2(X, y, penalty = c("MCP", "SCAD", "lasso"), gamma = switch(penalty,
  SCAD = 3.7, 3), alpha = 1, lambda.min = ifelse(n > p, 0.001, 0.05),
  nlambda = 100, lambda, eps = 1e-04, max.iter = 10000, convex = TRUE,
  dfmax = p, penalty.factor = rep(1, ncol(X)), warn = TRUE,
  returnX = FALSE, ...)
}
\description{
\code{ncvsurv2} is a modified function of ncvsurv (to resolve the bug from lamNames)
}
