% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestep.R
\name{PI.k.FUN}
\alias{PI.k.FUN}
\title{Inner function for one-step estimator}
\usage{
PI.k.FUN(tt, ebzi, xi, zi, k0 = 0, vi = NULL)
}
\description{
\code{PI.k.FUN} is an inner function to calculate score and negative information.
}
