% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestep.R
\name{Score.A.FUN}
\alias{Score.A.FUN}
\title{Calculate score and negative information (A) matrix for a Cox model}
\usage{
Score.A.FUN(data, betahat, rtn = "score")
}
\arguments{
\item{data}{a matrix with first column = U, second column = delta (0/1), rest = design}

\item{betahat}{at which coefficient level to evaluate score and negative information}

\item{rtn}{'Score' returns only the score function, 'Score+A' returns score and negative information (second derivative of PL), 'Score+A+Approx' returns score and negative information (approximated by -SS')}
}
\description{
\code{Score.A.FUN} evaluates score and negative information (A)
}
