% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestep.R
\name{U.fun}
\alias{U.fun}
\title{Calculate score for logistic regression.}
\usage{
U.fun(bet, dat)
}
\arguments{
\item{bet}{is the beta for evaluation}

\item{dat}{is the dataset}
}
\description{
\code{U.fun} evaluates scores for logistic regression.
}
